% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{check_D_U}
\alias{check_D_U}
\title{check_D_U}
\usage{
check_D_U(D_U_ctrl, ctrl_data, threshold)
}
\arguments{
\item{D_U_ctrl}{The list of the two matrices D and U for each genes.}

\item{ctrl_data}{The matrix of gene expression used to make D and U matrices.}

\item{threshold}{The threshold used to make D and U matrices. It's the proportion of genes that
must meet the conditions.}
}
\value{
This function returns an error message if D or U list doesn't meet the conditions. In normal case, the function returns an "OK" message.
}
\description{
This function checks if not exists genes both up and down expressed, and for each gene :
if the maximal down gene is less expressed in threshold% of cases,
if the minimal up gene is more expressed in threshold % of cases.
}
\examples{
# Example for check_D_U function
D_U = find_D_U_ctrl(ctrl_data, quant = 0.001, factor = 4, threshold = 0.99)
check_D_U(D_U_ctrl = D_U, ctrl_data, threshold = 0.9)
}
