% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_param.R
\name{choose_quantile}
\alias{choose_quantile}
\title{choose_quantile}
\usage{
choose_quantile(ctrl_data, simulation, FDR_goal, factor = 1.4,
  quantile_values = c(0.005, 0.01, 0.015, 0.02, 0.025, 0.03, 0.04, 0.05, 0.06,
  0.07, 0.08, 0.09, 0.1, 0.15, 0.2))
}
\arguments{
\item{ctrl_data}{A matrix with genes expressions in controls for all the patients.}

\item{simulation}{The list of initial data $initial_data and modified data in $simulated_data}

\item{FDR_goal}{The FDR you would like for this test.}

\item{factor}{The factor for the test The D limit will be quantmin/factor, and the U limit quantmax*factor.}

\item{quantile_values}{A vector with different values to test for the quantile.}
}
\value{
This function returns a matrix with 3 columns : the value of the quantile tested,
the FDR and the TPR of the test.
}
\description{
This function makes for each patient the naive test with multiple values for the quantile.
After that, it computes FDR and TPR, and print the quantile closest of your goal FDR.
}
\examples{
simulation = simplified_simulation(simu_data, fraction = 0.3, threshold = 60)
which_quantile = choose_quantile(ctrl_data, simulation)

}
