% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{complex_simulation}
\alias{complex_simulation}
\title{complex_simulation}
\usage{
complex_simulation(data)
}
\arguments{
\item{data}{vector or matrix of original data to edit.}
}
\value{
This function returns a list of three vectors or matrix :
initial_data, datas before the modification
simulated_data, datas with up and down modifications
changes_idx, index of datas with modifications
}
\description{
This function simulates the dysregulation (up and down) of datas, using the distribution of
dysregulation in real cancer datas.
}
