% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{compute_down_and_up_list}
\alias{compute_down_and_up_list}
\title{Compute D and U list in control samples.}
\usage{
compute_down_and_up_list(controls, threshold, s_max = 50)
}
\arguments{
\item{controls}{A matrix with the gene expressions for each patient.}

\item{threshold}{The proportion of expression that must be in the conditions.}

\item{s_max}{The maximum number of down and up-expressed gene for each genes.}
}
\value{
This function returns a list of two logical matrices :
- the D matrix, with TRUE if the row gene has a lower expression than the column gene,
- the U Matrix with TRUE if the row gene has a higher expression than the column gene.
}
\description{
For each gene, this function computes two lists. The D list, with genes less expressed in threshold% of controls,
and the U list with genes more expressed in threshold% of controls. These lists can be used in Penda test to compare
the gene rank in samples.
}
\examples{
# First, data are filtering to remove low expression values
controls = penda::data_ctrl[1:10, 1:10]
# Second, down and up list are computed
D_U_list = penda::compute_down_and_up_list(controls, threshold = 0.99, s_max = 50)
}
