% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_DU_cpp}
\alias{compute_DU_cpp}
\title{compute_DU_cpp}
\usage{
compute_DU_cpp(data_ctrl, threshold)
}
\arguments{
\item{data_ctrl}{A numeric matrix with the genes expressions for each patient.}

\item{threshold}{The proportion of genes that must be under or above the gene.}
}
\value{
This function returns a list of two logical matrices :
the D matrix, with TRUE if the row gene has a lower expression than the column gene,
and the U Matrix with TRUE if the row gene has a higher expression than the column gene.
And a string vector with the genes names.
}
\description{
This function makes the matrix D and U for all the genes.
}
