% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_DU_cpp_V2}
\alias{compute_DU_cpp_V2}
\title{compute_DU_cpp_V2}
\usage{
compute_DU_cpp_V2(data_ctrl, threshold, size_max)
}
\arguments{
\item{data_ctrl}{A numeric matrix with the genes expressions for each patient. Must be sort by the median.}

\item{threshold}{The proportion of genes that must be under or above the gene.}

\item{size_max}{The maximum number of down and up-expressed gene for each genes.}
}
\value{
This function returns a list of two logical matrices :
the D matrix, with the id of closest genes with a lower expression,
and the U Matrix with the id of closest genes with a higher expression.
And a string vector with the genes names.
}
\description{
This function makes the matrix D and U for all the genes with a new method.
}
