% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_DU_cppNA_V2}
\alias{compute_DU_cppNA_V2}
\title{compute_DU_cppNA}
\usage{
compute_DU_cppNA_V2(data_ctrl, threshold, nbNA, size_max)
}
\arguments{
\item{data_ctrl}{A numeric matrix with the genes expressions for each patient.}

\item{threshold}{The proportion of genes that must be under or above the gene.}

\item{nbNA}{The number of NA for each gene.}
}
\value{
This function returns a list of two logical matrices :
the D matrix, with TRUE if the row gene has a lower expression than the column gene,
and the U Matrix with TRUE if the row gene has a higher expression than the column gene.
And a string vector with the genes names.
}
\description{
This function makes the matrix D and U for all the genes without count NA values.
}
