% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_LH_cppNA}
\alias{compute_LH_cppNA}
\title{compute_LH_cppNA}
\usage{
compute_LH_cppNA(data_ctrl, threshold, nbNA, size_max)
}
\arguments{
\item{data_ctrl}{A numeric matrix with the genes expressions for each patient.}

\item{threshold}{The proportion of genes that must be under or above the gene.}

\item{nbNA}{The number of NA for each gene.}

\item{size_max}{The maximum number of L and H genes for each gene..}
}
\value{
This function returns a list of two logical matrices :
the L matrix, with the id of closest genes with a lower expression,
the H Matrix with the id of closest genes with a higher expression,
And the n string vector with the genes names.
}
\description{
This function makes the matrices L and H for all the genes without count NA values.
}
