% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{compute_lower_and_higher_lists}
\alias{compute_lower_and_higher_lists}
\title{Compute L and H list in control samples.}
\usage{
compute_lower_and_higher_lists(controls, threshold, s_max = 50)
}
\arguments{
\item{controls}{A matrix with the gene expressions for each patient.}

\item{threshold}{The proportion of expression that must be in the conditions.}

\item{s_max}{The maximum number of L and H genes for each gene.}
}
\value{
This function returns a list of two logical matrices :
- the L matrix, with TRUE if the row gene has a lower expression than the column gene,
- the H Matrix with TRUE if the row gene has a higher expression than the column gene.
}
\description{
For each gene, this function computes two lists.
The L list, with genes with a Lower expression in threshold% of controls,
and the H list with genes with an Higher expression in threshold% of controls.
These lists can be used in Penda test to compare the gene rank in samples.
}
\examples{
# First, data are ranking by median
controls = penda::penda_data_ctrl[1:10, 1:10]
median_gene = apply(controls, 1, median, na.rm = TRUE)
median_gene = sort(median_gene)
controls = controls[names(median_gene), ]
# Second, lower and higher lists are computed
L_H_list = penda::compute_lower_and_higher_lists(controls, threshold = 0.99, s_max = 50)
}
