% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version1ctrl.R
\name{compute_lower_and_higher_lists_1ctrl}
\alias{compute_lower_and_higher_lists_1ctrl}
\title{Compute L and H list in control sample.}
\usage{
compute_lower_and_higher_lists_1ctrl(control, s_max = 500)
}
\arguments{
\item{control}{A vector with the expressions in the control sample.}

\item{s_max}{The maximum number of L and H genes for each gene.}
}
\value{
This function returns a list of two numeric matrices :
- the L matrix, with for each row gene the id of the L genes in columns,
- the H Matrix, with for each row gene the id of the H genes in columns.
}
\description{
For each gene, this function computes two lists.
The L list, with s_max genes with the closer Lower expression in control,
and the H list with s_max genes with the closer Higher expression in control.
These lists can be used in Penda test to compare the gene rank in samples.
}
\examples{
# First, data are ranking by median
control = penda::penda_data_ctrl[1:10, 1]
control = sort(control)
# Second, lower and higher lists are computed
L_H_list = penda::compute_lower_and_higher_lists_1ctrl(control, s_max = 50)
}
