% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_case}
\alias{data_case}
\title{Transcriptome data of case samples.}
\format{An object of class \code{matrix} with 19177 rows and 44 columns.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
data_case
}
\description{
A matrix that contains normalized transcriptome data for case samples.
}
\keyword{datasets}
