% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ctrl}
\alias{data_ctrl}
\title{Transcriptome data of control samples.}
\format{An object of class \code{matrix} with 19177 rows and 44 columns.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
data_ctrl
}
\description{
A matrix that contains normalized transcriptome data for control samples.
}
\keyword{datasets}
