% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{detect_na_value}
\alias{detect_na_value}
\title{Find NA values}
\usage{
detect_na_value(controls, cancer_data, threshold, probes = TRUE)
}
\arguments{
\item{controls}{A matrix with datas to analyze.}

\item{cancer_data}{A matrix with other conditions datas to analyze.}

\item{threshold}{The maximum proportion of NA tolerated for each probe.}

\item{probes}{TRUE if you want to sort for probes, FALSE to sort for patients.}
}
\value{
This function returns a true false vector with true for the values to exclude.
}
\description{
This function detects probes with more than threshold percent of value undefined (NA).
}
\examples{
# Example for detect_zero_value function
data_ctrl = penda::data_ctrl
data_case = penda::data_case
null_values = detect_zero_value(data_ctrl, data_case, threshold = 0.99, min = 10)
data_ctrl_f = data_case[!null_values,]
data_case_f = data_case[!null_values,]
}
