% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{detect_zero_value}
\alias{detect_zero_value}
\title{detect_zero_value}
\usage{
detect_zero_value(ctrl_data, cancer_data, threshold)
}
\arguments{
\item{ctrl_data}{matrix with data to analyze}

\item{cancer_data}{matrix with other data to analyze}

\item{threshold}{maximum proportion of zero tolerated for each gene}
}
\value{
This function return a true false vector with true for the values to exclude.
}
\description{
This function detects genes with more than threshold percent of expression values equals to zero.
}
\examples{
null_values = detect_zero_value(ctrl_data, simu_data, threshold = 0.8)
ctrl_data = ctrl_data[!null_values,]
simu_data = simu_data[!null_values,]
}
