% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{draw_results}
\alias{draw_results}
\title{draw_results}
\usage{
draw_results(results)
}
\arguments{
\item{results}{The list TP, FP, FN and TN return by "results_simulation"}
}
\value{
This function returns a barplot of results, FDR and TPR.
}
\description{
This function computes the the FDR and the TPR and makes a ggplot of the results.
}
