% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{dysreg_simulation}
\alias{dysreg_simulation}
\title{dysreg_simulation}
\usage{
dysreg_simulation(simu_data, fraction, threshold)
}
\arguments{
\item{simu_data}{vector of original data to be modified}

\item{fraction}{fraction of data to modify}

\item{threshold}{under this threshold, data modification is +-15. Above, it is *2 or /2.}
}
\value{
This function returns a vector of modified data which simulates up and down regulation.
}
\description{
This function simulates the dysregulation (up and down) of datas.
}
\examples{
dysreg_simulation(simu_data, fraction = 0.3, threshold = 60)

}
