% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{find_D_U_ctrl_size}
\alias{find_D_U_ctrl_size}
\title{find_D_U_ctrl_size}
\usage{
find_D_U_ctrl_size(ctrl_data, threshold, s_max = 50)
}
\arguments{
\item{ctrl_data}{A matrix with the gene expressions for each patient.}

\item{threshold}{The proportion of expression that must be in the conditions.}

\item{s_max}{The maximal number of down and up-expressed gene for each genes.}
}
\value{
This function returns a list of two logical matrices :
the D matrix, with TRUE if the row gene has a lower expression than the column gene,
and the U Matrix with TRUE if the row gene has a higher expression than the column gene.
}
\description{
This function ranks each gene and finds the genes which are more or less exprimed.
If there are too many genes, the closest are kept.
}
