% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{get_random_variable}
\alias{get_random_variable}
\title{get_random_variable}
\usage{
get_random_variable(F, x, nb_r, ...)
}
\arguments{
\item{F}{The density function.}

\item{x}{A vector with the definition area of the function.}

\item{nb_r}{The number of random values wanted}
}
\value{
This function returns a vector with nb_r values of F.
}
\description{
This function gives a random values following a given density function.
}
