% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{group_genes}
\alias{group_genes}
\title{Group genes with similar expression}
\usage{
group_genes(controls, cancer_data, size_grp = 100, quant = 0.05)
}
\arguments{
\item{controls}{The matrix of gene expression in controls.}

\item{cancer_data}{The matrix of gene expression in cases (ex: cancer).}

\item{size_grp}{The size of the groups of genes.}

\item{quant}{The quantile of gene expression in control. Cancer genes outside this limit are considered dysregulated.}
}
\value{
This function returns a matrix of four columns :
Limit min, the first value of the group,
Limit max, the last value of the group,
Proportion of dysregulation, the proportion of cancer genes outside the quantile limit.
Deltas dysregul, a list with all the value of cancer genes - control genes.
}
\description{
This function makes group of genes with similar expression. For each group,
the proportion of dysregulated genes in cancer and the extent of variation are computed.
}
