% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{Heaviside}
\alias{Heaviside}
\title{Heaviside}
\usage{
Heaviside(x, step = 0)
}
\arguments{
\item{x}{The number to evaluate}

\item{step}{The limit between heaviside = 0 and heaviside = 1}
}
\value{
This function returns 0 if the argument is under the step, 1 if not.
}
\description{
The heaviside function value is zero for argument under the step, and one for argument above.
}
\examples{
Heaviside(1+98-103, step = 0)

}
