% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{make_dataset}
\alias{make_dataset}
\title{Making the Penda Dataset}
\usage{
make_dataset(controls, cancer_data, detectlowvalue = TRUE,
  detectNA = TRUE, threshold = 0.99, val_min = NA)
}
\arguments{
\item{controls}{A matrix with datas to analyze.}

\item{cancer_data}{A matrix with other conditions datas to analyze.}

\item{detectlowvalue}{If detectlowvalue is true, we apply the function "detect_zero_value".}

\item{detectNA}{If detectnavalue is true, we apply the function "detect_na_value" on probes and on patients.}

\item{threshold}{The maximum proportion of expression under val_min or NA tolerated for each gene.}

\item{val_min}{The minimum value accepted. If val_min is NA, we compute this value with mixtools.}
}
\value{
This function return a list with data_ctrl and data_case.
}
\description{
This function makes a dataset with controls and cancer_data pre-filtred and sorted by median in controls.
}
\examples{
# Example for detect_zero_value function
data_ctrl = penda::penda_data_ctrl
data_case = penda::penda_data_case
dataset = penda::make_dataset(controls = data_ctrl,
                              cancer_data = data_case,
                              detectlowvalue = TRUE,
                              detectNA = TRUE,
                              threshold = 0.99,
                              val_min = NA)
data_ctrl = dataset$data_ctrl
data_case = dataset$data_case
}
