% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version1ctrl.R
\name{make_dataset_1ctrl}
\alias{make_dataset_1ctrl}
\title{Making the Penda Dataset for one control samples}
\usage{
make_dataset_1ctrl(control, data_case)
}
\arguments{
\item{data_case}{The matrix with datas to analyze (ex: tumors samples).}

\item{controls}{The vector with reference data to analyze (ex: control sample).}
}
\value{
This function return a list with preprocessed data_ctrl and data_case, and
the vector `info` with the different parameters.
}
\description{
This function makes the Penda dataset with control and cases
sorted by value in control.
}
\examples{
# Example for make_dataset_1ctrl function
data_ctrl = penda::penda_data_ctrl[1:10, 1]
data_case = penda::penda_data_case[1:10, 1:3]
dataset = penda::make_dataset_1ctrl(control = data_ctrl,
                                    data_case = data_case)
data_ctrl = dataset$data_ctrl
data_case = dataset$data_case
}
