% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{multiple_tests}
\alias{multiple_tests}
\title{multiple_tests}
\usage{
multiple_tests(ctrl_data, D_U_ctrl, iterations, simulation, multiple_values,
  unique_value, threshold_change = TRUE)
}
\arguments{
\item{ctrl_data}{A matrix with genes expressions in controls for all the patients.}

\item{D_U_ctrl}{The list of Down and Up-expressed genes matrices in the control.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list
no longer changes, iterations are stop before.}

\item{simulation}{The list of initial data $initial_data and modified data in $simulated_data}

\item{multiple_values}{A vector with different values to test for the chosen parameter (threshold or quantile).}

\item{unique_value}{The value of the parameter that does not vary.}

\item{threshold_change}{TRUE if you want to vary the threshold, FALSE if you want to vary the quantile.}
}
\value{
This function returns a matrix with : the patient, the value of the changing paramter,
the FDR and the TPR of the test.
}
\description{
This function makes for each patient the test with multiple values for the treshold or the quantile.
After that, it computes FDR and TPR.
}
