% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{p1}
\alias{p1}
\title{P1}
\usage{
p1(x)
}
\arguments{
\item{x}{}
}
\value{
This function returns
}
\description{
P1 models the density function of the expression multiplicative factor for the
up-regulation of a gene with an expression above 100.
}
