% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{p2}
\alias{p2}
\title{P2}
\usage{
p2(x)
}
\arguments{
\item{x}{vector of input values}
}
\value{
This function returns the value of p2 for x (between 0 and 1).
}
\description{
P2 models the density function of the expression divisor for the
down-regulation of a gene with an expression above 100.
}
