% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{p3}
\alias{p3}
\title{P3}
\usage{
p3(x)
}
\arguments{
\item{x}{vector of input values}
}
\value{
This function returns the value of p3 for x (between 0 and 1).
}
\description{
P2 models the density function of the expression modifier (to add or substract) for the
dysregulation of a gene with an expression under 100.
}
