% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{patient_test}
\alias{patient_test}
\title{patient_test}
\usage{
patient_test(ctrl_data, patient_genes, iterations, D_U_ctrl, threshold,
  quant_test = 0, factor_test = 1)
}
\arguments{
\item{ctrl_data}{A matrix with genes expressions in controls for all the patients.}

\item{patient_genes}{A vector with all the genes expressions for one patient.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list
no longer changes, iterations are stop before.}

\item{D_U_ctrl}{The list of Down and Up-expressed genes matrices in the control.}

\item{threshold}{The threshold for regulation_test}

\item{quant_test}{The quantile for the test. When D and U lists are void, we use the naive method.}

\item{factor_test}{The factor for the test. The limit D will be quant(gene)/factor, and the limit U quant(gene)*factor.}
}
\value{
This function return a list with two vectors :
D, with TRUE for genes down-regulated
U, with TRUE for genes up-regulated.
}
\description{
This function makes the test of the dysregulation of all genes for one patient.
}
\examples{
simulation = ctrl_data[,ncol(ctrl_data)-3 : ncol(ctrl_data)]
simulation = simplified_simulation(simulation, fraction = 0.3, threshold = 60)
ctrl_data = ctrl_data[,-ncol(ctrl_data)-3:ncol(ctrl_data)]
D_U_ctrl = find_D_U_ctrl(ctrl_data, quant = 0.001, factor = 4, threshold = 0.99)
test_patient1 = patient_test(ctrl_data, simulation$simulated_data[,1], iterations = 10, D_U_ctrl, threshold = 0.2)

}
