% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penda_data_case}
\alias{penda_data_case}
\title{Expression matrix of cases}
\format{
A matrix with 3000 rows and 44 columns
}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
penda_data_case
}
\description{
A matrix...
}
\keyword{datasets}
