% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version1ctrl.R
\name{penda_test_1ctrl}
\alias{penda_test_1ctrl}
\title{Test gene expression variations in a serie of samples}
\usage{
penda_test_1ctrl(samples, iterations, L_H_list, threshold)
}
\arguments{
\item{samples}{A vector or a matrix with all the genes expressions for each sample.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list no longer changes, iterations are stop before.}

\item{L_H_list}{The list of lower and higher expressed genes matrices in the control.}

\item{threshold}{If Ll/Lh and Hh/Hl are under this threshold, the expression not change.}
}
\value{
This function return a list of two matrices (down_genes and up_genes)
}
\description{
This function applies the function sample_test_1ctrl to a serie of samples.
It's a special version of the Penda method for the datasets with only 1 control,
if you want the normal Penda method used penda::penda_test
}
\examples{
# First, load and define the data
control = penda::penda_data_ctrl[1:10, 1]
samples = penda::penda_data_case[1:10, 1:3]
# Second, lower and higher lists are computed
L_H_list = penda::compute_lower_and_higher_lists_1ctrl(control, s_max = 50)
# Third, test is the expression is deregulated for each gene of a sample
res_penda = penda::penda_test_1ctrl(samples,
                              iterations = 20,
                              L_H_list = L_H_list,
                              threshold = 0.03)
}
