% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{quantile_test}
\alias{quantile_test}
\title{Simple test of gene expression variations in a serie of samples}
\usage{
quantile_test(controls, samples, quant, factor)
}
\arguments{
\item{controls}{A matrix with genes expressions in controls for all the patients.}

\item{samples}{A vector or a matrix with all the genes expressions for each sample.}

\item{quant}{The quantile of the control genes expression, quantile(c(quant, 1-quant)).}

\item{factor}{The factor for the quantile. The down-regulated limit will be quantmin/factor, and the up-regulated limit quantmax*factor.}
}
\value{
This function return a list with two vectors or matrix :
D, with TRUE for genes down-regulated
U, with TRUE for genes up-regulated.
}
\description{
This function makes the quantile test of the dysregulation of all genes.
The patient's gene is compared at the gene expression in controls sample.
}
\examples{
# First, load and define the data
controls = penda::penda_data_ctrl[1:10, 1:10]
samples = penda::penda_data_case[1:10, 1:3]
# Second, test is the expression is deregulated for each gene of a sample
res_quantile = penda::quantile_test(controls,
                                samples,
                                quant = 0.03,
                                factor = 1.4)
}
