% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version1ctrl.R
\name{regulation_test_1ctrl}
\alias{regulation_test_1ctrl}
\title{Test gene expression variation for a given gene}
\usage{
regulation_test_1ctrl(gene, L_H_list, sample, threshold)
}
\arguments{
\item{gene}{The name of the gene to analyze.}

\item{L_H_list}{The list of lower and higher expressed genes matrices in the control.}

\item{sample}{The vector of genes expressions for one patient.}

\item{threshold}{If Ll/Lh and Hh/Hl are under this threshold, the expression not change.}
}
\value{
This function returns 0 if the gene expression has not changed,
1 if the gene is up-regulated and -1 if the gene is down-regulated.
}
\description{
This function makes the test of expression dysregulation of one gene for a
patient compared to control. It's an adapted method for the datasets with only
1 control which use only the rank method. If you want the normal Penda method used penda::regulation_test
}
\examples{
# First, load and define the data
dataset = penda::make_dataset_1ctrl(
  penda::penda_data_ctrl[1:10, 1],
  penda::penda_data_case[1:10, 1:3]
)
controls = dataset$data_ctrl
sample = dataset$data_case[,1]
gene = names(sample)[1]
# Second, lower and higher list are computed
L_H_list = penda::compute_lower_and_higher_lists_1ctrl(controls, s_max = 50)
# Third, test is the expression is deregulated for a given gene.

expression = penda::regulation_test_1ctrl(gene,
                                    L_H_list,
                                    sample,
                                    threshold = 0.03)
}
