% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version1ctrl.R
\name{sample_test_1ctrl}
\alias{sample_test_1ctrl}
\title{Test gene expression variations for a given sample}
\usage{
sample_test_1ctrl(sample, iterations, threshold)
}
\arguments{
\item{sample}{A vector with all the genes expressions for one patient.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list no longer changes, iterations are stop before.}

\item{threshold}{The threshold for regulation_test.}
}
\value{
This function return a list with two vectors :
D, with TRUE for genes down-regulated
U, with TRUE for genes up-regulated.
}
\description{
This function makes the test of the dysregulation of all genes for one patient.
It's the adapted method for the datasets with only 1 control,
if you want the normal Penda method used penda::sample_test
}
\examples{
# First, load and define the data
dataset = penda::make_dataset_1ctrl(
  penda::penda_data_ctrl[1:10, 1],
  penda::penda_data_case[1:10, 1:3]
)
controls = dataset$data_ctrl
sample = dataset$data_case[,1]
# Second, lower and higher lists are computed
L_H_list = penda::compute_lower_and_higher_lists_1ctrl(controls, s_max = 50)
# Third, test is the expression is deregulated for each gene of a sample.

res_sample = penda::sample_test_1ctrl(sample,
                                iterations = 20,
                                threshold = 0.03)
}
