% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_param.R
\name{select_threshold_param}
\alias{select_threshold_param}
\title{Select the better threshold to reach a given FDR.}
\usage{
select_threshold_param(which_threshold, FDR_max = 0.05)
}
\arguments{
\item{which_threshold}{The matrix result of choose_threshold, with patient, FDR and threshold.}

\item{FDR_max}{The maximum FDR wanted.}
}
\value{
This function returns a list of 2 parameters : the best threshold and the associated FDR.
}
\description{
This function use the result of choose_threshold to give the parameter which allows to reach
the given FDR. If there are multiple patients, it uses the median
}
\examples{
# First, load and define the data, make the simulation
controls = penda::data_ctrl[1:10, 1:10]
simu_data = penda::data_ctrl[1:10,11:16]
simulation = penda::simplified_simulation(simu_data, fraction = 0.3, threshold = 60)
# Second, down and up list are computed
D_U_list = penda::compute_down_and_up_list(controls, threshold = 0.99, s_max = 50)
# Third, choose the vector of threshold to test
threshold_values = c(0.1, 0.2, 0.3, 0.4, 0.5)
# Fourth, make the test on simulation for different thresholds
which_threshold = penda::choose_threshold(controls,
                                          D_U_list,
                                          iterations = 20,
                                          simulation,
                                          threshold_values,
                                          quant_test = 0,
                                          factor_test = 1)
# Fifth, use the function to automatically choose the better threshold
best_threshold = penda::select_threshold_param(which_threshold, FDR_goal = 0.05)
}
