% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{step0}
\alias{step0}
\title{step0}
\usage{
step0(ctrl_data, patient_genes, quant)
}
\arguments{
\item{ctrl_data}{The matrix with genes expressions in controls for all the patients.}

\item{patient_genes}{The vector of genes expressions for one patient.}

\item{quant}{The quantile for the expression distribution limit.}
}
\value{
This function return a logical vector with TRUE for genes outside of the distribution limit.
}
\description{
This function makes the list of genes at the extremities of expression distribution.
}
\examples{
simulation = ctrl_data[,ncol(ctrl_data)-3 : ncol(ctrl_data)]
simulation = simplified_simulation(simulation, fraction = 0.3, threshold = 60)
ctrl_data = ctrl_data[,-ncol(ctrl_data)-3:ncol(ctrl_data)]
step0(ctrl_data, simulation$simulated_data[,1], quant = 0.03)

}
