% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_param.R
\name{choose_threshold}
\alias{choose_threshold}
\title{choose_threshold}
\usage{
choose_threshold(ctrl_data, D_U_ctrl, iterations, simulation, threshold_values,
  FDR_goal = 0.06, quant_test = 0, factor_test = 1)
}
\arguments{
\item{ctrl_data}{A matrix with genes expressions in controls for all the patients.}

\item{D_U_ctrl}{The list of Down and Up-expressed genes matrices in the control.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list
no longer changes, iterations are stop before.}

\item{simulation}{The list of initial data $initial_data and modified data in $simulated_data}

\item{threshold_values}{A vector with different values to test for the threshold.}

\item{FDR_goal}{The FDR you would like for this test.}

\item{quant_test}{The quantile for the test. When D and U lists are void, we use the naive method.}

\item{factor_test}{The factor for the test. The limit D will be quant(gene)/factor, and the limit U quant(gene)*factor.}
}
\value{
This function returns a matrix with 4 columns : the patient number, the value of the threshold tested,
the FDR and the TPR of the test.
}
\description{
This function makes for each patient the test with multiple values for the treshold.
After that, it computes FDR and TPR, and print the threshold closest of your goal FDR.
}
