% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{complex_simulation}
\alias{complex_simulation}
\title{Simulation of dysregulation that takes in account real datas}
\usage{
complex_simulation(controls, cancer_data, data, size_grp = 100, quant = 0.05)
}
\arguments{
\item{controls}{A matrix with genes expressions in controls for all the patients.}

\item{cancer_data}{A matrix with dysregulated genes expressions for all the patients.}

\item{data}{vector or matrix of original data to edit.}

\item{size_grp}{The size of each group of genes for the grouping.}

\item{quant}{The quantile of gene expression for the grouping.}
}
\value{
This function returns a list of three vectors or matrices :
initial_data, datas before the modification
simulated_data, datas with up and down modifications
changes_idx, index of datas with modifications
}
\description{
This function simulated the dysregulation of datas, using the real distribution of genes, proportion of
dysregulation and difference between control and cancer.
}
\examples{
# First, load and define the data
controls = penda::penda_data_ctrl[1:1000, 1:10]
samples = penda::penda_data_case[1:1000, 1:3]
simu_data = penda::penda_data_ctrl[1:1000,11:16]
# Second, make the simulation
simulation = penda::complex_simulation(controls,
                                       samples,
                                       simu_data,
                                       size_grp = 80,
                                       quant = 0.05)
}
