% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{compute_down_and_up_list}
\alias{compute_down_and_up_list}
\title{compute_down_and_up_list}
\usage{
compute_down_and_up_list(controls, threshold, s_max = 50)
}
\arguments{
\item{controls}{A matrix with the gene expressions for each patient.}

\item{threshold}{The proportion of expression that must be in the conditions.}

\item{s_max}{The maximum number of down and up-expressed gene for each genes.}
}
\value{
This function returns a list of two logical matrices :
the D matrix, with TRUE if the row gene has a lower expression than the column gene,
and the U Matrix with TRUE if the row gene has a higher expression than the column gene.
}
\description{
This function ranks each gene and finds the genes which are more or less exprimed.
It's the same than "find_D_U_ctrl_size" but faster because uses rcpp to rank.
}
\examples{
# First, data are filtering to remove low expression values
controls = penda::data_ctrl[1:10, 1:10]
# Second, down and up list are computed
D_U_list = penda::compute_down_and_up_list(controls, threshold = 0.99, s_max = 50)
}
