% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{detect_na_value}
\alias{detect_na_value}
\title{Find NA values}
\usage{
detect_na_value(controls, data_case, threshold, probes = TRUE)
}
\arguments{
\item{controls}{The first matrix with datas to analyze (ex: controls samples).}

\item{data_case}{The second matrix with datas to analyze (ex: tumors samples).}

\item{threshold}{The maximum proportion of NA tolerated for each probe (or sample).}

\item{probes}{TRUE if you want to analyze probes, FALSE to analyze samples.}
}
\value{
This function returns a True/False vector with True for the values to exclude.
}
\description{
This function detects probes or sample with more than threshold percent of
value undefined (NA).
}
\examples{
# Example for detect_na_value function
data_ctrl = penda::penda_data_ctrl
data_case = penda::penda_data_case
na_values = penda::detect_na_value(data_ctrl, data_case, threshold = 0.99)
data_ctrl_f = data_case[!na_values,]
data_case_f = data_case[!na_values,]
}
