% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{detect_zero_value}
\alias{detect_zero_value}
\title{detect_zero_value}
\usage{
detect_zero_value(ctrl_data, cancer_data, threshold, min = 0)
}
\arguments{
\item{ctrl_data}{A matrix with data to analyze.}

\item{cancer_data}{A matrix with other data to analyze.}

\item{threshold}{The maximum proportion of expression under min tolerated for each gene.}

\item{min}{The minimum value accepted.}
}
\value{
This function returns a true false vector with true for the values to exclude.
}
\description{
This function detects genes with more than threshold percent of expression values under the min value.
}
