% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{detect_zero_value}
\alias{detect_zero_value}
\title{detect_zero_value}
\usage{
detect_zero_value(controls, cancer_data, threshold, min = 0)
}
\arguments{
\item{controls}{A matrix with datas to analyze.}

\item{cancer_data}{A matrix with other conditions datas to analyze.}

\item{threshold}{The maximum proportion of expression under min tolerated for each gene.}

\item{min}{The minimum value accepted.}
}
\value{
This function returns a true false vector with true for the values to exclude.
}
\description{
This function detects genes with more than threshold percent of expression values under the min value.
NA expression values are not considered.
}
\examples{
# Example for detect_zero_value function
data_ctrl = penda::data_ctrl
data_case = penda::data_case
null_values = detect_zero_value(data_ctrl, data_case, threshold = 0.99, min = 10)
data_ctrl_f = data_case[!null_values,]
data_case_f = data_case[!null_values,]
}
