% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{detect_zero_value}
\alias{detect_zero_value}
\title{Find low values}
\usage{
detect_zero_value(controls, data_case, threshold, min = 0)
}
\arguments{
\item{controls}{The first matrix with datas to analyze (ex: controls samples).}

\item{data_case}{The second matrix with datas to analyze (ex: tumors samples).}

\item{threshold}{The maximum proportion of expression under `min` tolerated for each gene.}

\item{min}{The minimum value accepted.}
}
\value{
This function returns a True/False vector with True for the values to exclude.
}
\description{
This function detects genes with more than `threshold` percent of expression
values under the min value. NA expression values are not considered.
}
\examples{
# Example for detect_zero_value function
data_ctrl = penda::penda_data_ctrl
data_case = penda::penda_data_case
null_values = penda::detect_zero_value(data_ctrl, data_case, threshold = 0.99, min = 10)
data_ctrl_f = data_case[!null_values,]
data_case_f = data_case[!null_values,]
}
