% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{dysreg_simulation}
\alias{dysreg_simulation}
\title{dysreg_simulation}
\usage{
dysreg_simulation(data, fraction, threshold)
}
\arguments{
\item{data}{vector or matrix of original data to edit}

\item{fraction}{fraction of data to modify}

\item{threshold}{under this threshold, data modification is +-15. Above, it is *2 or /2.}
}
\value{
This function returns a list of three vectors or matrix :
initial_data, datas before the modification
simulated_data, datas with up and down modifications
changes_index, index of datas with modifications
}
\description{
This function simulates the dysregulation (up and down) of datas.
}
\examples{
dysreg_simulation(simu_data, fraction = 0.3, threshold = 60)

}
