% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{get_random_variable}
\alias{get_random_variable}
\title{get_random_variable}
\usage{
get_random_variable(F, x, nb_r, ...)
}
\arguments{
\item{F}{The density function.}

\item{x}{A vector with values of the function.}

\item{nb_r}{The number of random values wanted}
}
\value{
This function returns a vector with nb_r values of F.
}
\description{
This function allows to get random values on a density function.
}
\examples{
x = seq(-20, 20, 0.05)
F = function(x, u, a){
 (1/(a*sqrt(2*pi))) * exp(-(((x-u)^2)/(2*a^2)))
}
get_random_variable(F, x, nb_r = 10000, u = 0, a = 1)

}
