% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{make_dataset}
\alias{make_dataset}
\title{Making the Penda Dataset}
\usage{
make_dataset(
  controls,
  data_case,
  detectlowvalue = TRUE,
  detectNA = TRUE,
  threshold = 0.99,
  val_min = NA,
  bimod = TRUE
)
}
\arguments{
\item{controls}{The first matrix with datas to analyze (ex: controls samples).}

\item{data_case}{The second matrix with datas to analyze (ex: tumors samples).}

\item{detectlowvalue}{If detectlowvalue is True, genes with a low values in more than `threshold*100` \% of samples are removed.}

\item{detectNA}{If detectNA is True, genes and samples with more than `threshold*100` \% of NA values are removed.}

\item{threshold}{The maximum proportion of expression under val_min or NA tolerated for each gene or sample.}

\item{val_min}{The minimum value accepted. If val_min is NA, we compute this value as specified by the parameter `bimod`.}

\item{bimod}{If bimod is True and val_min NA, val_min is computed by mixtools::normalmixEM for a bimodal distribution, to search the value of the first peak. If bimod is False and val_min NA, val_min is computed for an unimodal distribution with a quantile of 0.1.}
}
\value{
This function return a list with preprocessed data_ctrl and data_case, and
the vector `info` with the different parameters.
}
\description{
This function makes the Penda dataset with controls and cases pre-filtred and
sorted by median in controls.
}
\examples{
# Example for make_dataset function
data_ctrl = penda::penda_data_ctrl
data_case = penda::penda_data_case
dataset = penda::make_dataset(controls = data_ctrl,
                              data_case = data_case,
                              detectlowvalue = TRUE,
                              detectNA = TRUE,
                              threshold = 0.99,
                              val_min = NA,
                              bimod = TRUE)
data_ctrl = dataset$data_ctrl
data_case = dataset$data_case
}
