% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{original_test}
\alias{original_test}
\title{original_test}
\usage{
original_test(gene, D_U_ctrl, patient_genes, threshold, ctrl_data, quant = 0)
}
\arguments{
\item{gene}{The name of the gene to analyze.}

\item{D_U_ctrl}{The list of Down and Up-expressed genes matrices in the control.}

\item{patient_genes}{The vector of genes expressions for one patient.}

\item{threshold}{If Dd/Du and Uu/ud are under this threshold, the expression not change}

\item{ctrl_data}{The matrix with genes expressions in controls for all the patients.}
}
\value{
This function returns 0 if the gene expression has not changed,
1 if the gene is up-regulated and -1 if the gene is down-regulated.
}
\description{
Original_test is the older version of regulation_test. In this version, we don't use the
quantile method in case of void down or up-expressed genes list.
}
