% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{patient_test}
\alias{patient_test}
\title{patient_test}
\usage{
patient_test(ctrl_data, patient_genes, quant_0, iterations, D_U_ctrl, threshold)
}
\arguments{
\item{ctrl_data}{A matrix with genes expressions in controls for all the patients.}

\item{patient_genes}{A vector with all the genes expressions for one patient.}

\item{quant_0}{The quantile of the expression distribution for the step_0.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list
no longer changes, iterations are stop before.}

\item{D_U_ctrl}{The list of Down and Up-expressed genes matrix in the control.}

\item{threshold}{The threshold for regulation_test}
}
\value{
This function return a list with two vector :
D, with TRUE for genes down-regulated
U, with TRUE for genes up-regulated.
}
\description{
This function makes the test of the dysregulation of all genes for one patient.
}
