% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{regulation_test}
\alias{regulation_test}
\title{Test gene expression variation for a given gene}
\usage{
regulation_test(gene, L_H_list, sample, threshold)
}
\arguments{
\item{gene}{The name of the gene to analyze.}

\item{L_H_list}{The list of lower and higher expressed genes matrices in the control.}

\item{sample}{The vector of genes expressions for one patient.}

\item{threshold}{If Ll/Lh and Hh/Hl are under this threshold, the expression not change.}
}
\value{
This function returns 0 if the gene expression has not changed,
1 if the gene is up-regulated and -1 if the gene is down-regulated.
}
\description{
This function makes the test of expression dysregulation of one gene for a
patient compared to control. It's an hybrid method, which use the rank method
and the quantile method in addition when there is no lower or higher list.
}
\examples{
examples/ex_regulation_test.R

}
