% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{results_simulation}
\alias{results_simulation}
\title{results_simulation}
\usage{
results_simulation(D_matrix, U_matrix, simulation)
}
\arguments{
\item{D_matrix}{The matrix of down-regulated genes.}

\item{U_matrix}{The matrix of up-regulated genes.}

\item{simulation}{The list of initial data $initial_data and modified data in $simulated_data}
}
\value{
This function returns a list of four integer :
TP, the number of true postive results
FP, the number of false postive results
FN, the numbler of false negative results
TN, the numbler of true negative results
}
\description{
This function computes the number of false positive, true positive, false negative
and false positive results after the test on a simulation.
}
