% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{step0}
\alias{step0}
\title{step0}
\usage{
step0(ctrl_data, patient_genes, quant)
}
\arguments{
\item{ctrl_data}{The matrix with genes expressions in controls for all the patients.}

\item{patient_genes}{The vector of genes expressions for one patient.}

\item{quant}{The quantile for the expression distribution limit.}
}
\value{
This function return a logical vector with TRUE for genes outside of the distribution limit.
}
\description{
This function makes the list of genes at the extremities of expression distribution.
}
