% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{step0}
\alias{step0}
\title{step0}
\usage{
step0(genes_data, patient_genes, quant)
}
\arguments{
\item{genes_data}{a matrix with genes expressions in controls for all the patients}

\item{patient_genes}{all the genes expression for one patient}

\item{quant}{the quantile for the expression distribution}
}
\value{
This function return a vector with all the genes outside of the distribution limit
}
\description{
This function makes the list of genes at the extremities of expression distribution.
}
\examples{
data_regulate = ctrl_data[,ncol(ctrl_data)]
data_regulate = dysreg_simulation(data_regulates, fraction = 0.3, threshold = 60)
ctrl_data = ctrl_data[,-ncol(ctrl_data)]
step0(ctrl_data, data_regulate, quant = 0.05)

}
