% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_param.R
\name{test_multiple_thresholds}
\alias{test_multiple_thresholds}
\title{Test gene expression variations for multiple thresholds}
\usage{
test_multiple_thresholds(simulation, threshold_values, iterations)
}
\arguments{
\item{simulation}{The simulation with the list of initial data $initial_data and modified data in $simulated_data.}

\item{threshold_values}{The vector of values of the threshold for the regulation test.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list
no longer changes, iterations are stop before.}
}
\value{
This function returns a list of two logical matrices. The D matrix with for each threshold
TRUE for genes down-regulated, and the U matrix with TRUE for genes up-regulated.
}
\description{
This function makes the Penda dysregulation test for different values of the threshold.
}
\examples{
examples/ex_test_multiple_thresholds.R

}
