% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{test_multiple_thresholds}
\alias{test_multiple_thresholds}
\title{test_multiple_thresholds}
\usage{
test_multiple_thresholds(ctrl_data, D_U_ctrl, simulation, threshold_values,
  iterations, quantile)
}
\arguments{
\item{ctrl_data}{A matrix with genes expressions in controls for all the patients.}

\item{D_U_ctrl}{The list of Down and Up-expressed genes matrices in the control.}

\item{simulation}{The list of initial data $initial_data and modified data in $simulated_data}

\item{threshold_values}{The vector of values of the threshold for regulation_test.}

\item{iterations}{The maximal number of iterations for the test. If the dysregulation list
no longer changes, iterations are stop before.}

\item{quantile}{The quantile of the expression distribution for the step_0.}
}
\value{
This function returns a list of two logical matrices. The D matrix with for each threshold
TRUE for genes down-regulated, and the U matrix with TRUE for genes up-regulated.
}
\description{
This function makes patient_test for different values of the threshold.
}
